+++ docs/Changelog.md
@@ -2,8 +2,8 @@
 
 ## 3.0.0 (TBA)
 - Implemented OIDC activation
-- Added PreApproval Screens & PreApprovalScreenRecorder support ([documentation](Using-Operations.md#pre-approval-screens))
-- Added `mobileTokenData` to reject request for passing customer-specific data
+- Added PreApproval Screens & PreApprovalScreensRecorder support ([documentation](Using-Operations.md#pre-approval-screens))
+- Added `reject(operation, reason)` overload that includes `operation.mobileTokenData` in the request for passing customer-specific data
 
 ## 2.1.0
 - Added `mobileTokenData` to authorize request for passing customer-specific data ([documentation](Using-Operations.md#Passing-Additional-Mobile-Token-Data))